LowWatermark refactored, interface extracted, added dummy implementation.
